feat: abstracted gizmo drawing functions
